cctv_access_tracking_django_service/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-timescaledb==0.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
